+++ test/unit/test_integer.cpp
@@ -1,0 +1,46 @@
+// Copyright 2024, reasuke.
+
+#include "gtest/gtest.h"
+#include "unit_test_utils.hpp"
+
+extern "C" {
+#include "libft.h"
+}
+
+TEST(ft_abs, test) {
+  EXPECT_EQ(ft_abs(42), 42);
+  EXPECT_EQ(ft_abs(0), 0);
+  EXPECT_EQ(ft_abs(-42), 42);
+}
+
+TEST(ft_chmax, chooseMax) {
+  int a = 0;
+  int b = 42;
+
+  EXPECT_TRUE(ft_chmax(&a, b));
+  EXPECT_EQ(a, 42);
+}
+
+TEST(ft_chmax, notChooseMax) {
+  int a = 42;
+  int b = 0;
+
+  EXPECT_FALSE(ft_chmax(&a, b));
+  EXPECT_EQ(a, 42);
+}
+
+TEST(ft_chmin, chooseMin) {
+  int a = 42;
+  int b = 0;
+
+  EXPECT_TRUE(ft_chmin(&a, b));
+  EXPECT_EQ(a, b);
+}
+
+TEST(ft_chmin, notChooseMin) {
+  int a = 0;
+  int b = 42;
+
+  EXPECT_FALSE(ft_chmin(&a, b));
+  EXPECT_EQ(a, 0);
+}
